refactor: noctaCrdt 라이브러리 구조변경
- src파일에 주요 파일들 이동
- Interface의경우 types로 따로이동
